feat: add form controls for account and category
Add form controls for filtering by accounts and categories.

I was not able to get the filters to work for the date range or sorting
the date so I just commented them out with a TODO.

This was implemented rather hastily so I have every intention of
refactoring this at a later date.